Refactor unnecessary views away
